config: split config.go into multiple files
ZoneConfigs and the SystemConfig are different beasts; give them each
their own file. Split out key encoding and decoding routines into their
own file, too, which will soon be responsible for constructing keys in
system.zones.
